requirements.txt
pandas
xlrd

transliterate
python-Levenshtein
pymorphy2
nltk
scikit-learn
